pyproject.toml
[tool.poetry]
name = "frisbee-ai"
version = "0.1.0"
description = ""
authors = ["Caleb <calebcauthon@gmail.com>"]
readme = "README.md"
packages = [{include = "frisbee_ai"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12.0"
roboflow = "^1.1.7"
pillow = "^10.1.0"
supervision = "^0.16.0"
opencv-python = "^4.8.1.78"
ultralytics = "^8.0.202"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
